# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gmail Unsubscribe Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/events/actions/{plan_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Action Events",
        "operationId": "action_events_events_actions__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Replayable action-state event stream",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/action-plans": {
      "post": {
        "tags": [
          "action-plans"
        ],
        "summary": "Create Plan",
        "operationId": "create_plan_api_action_plans_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "unsubscribe_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unsubscribe Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/action-plans/{plan_id}": {
      "get": {
        "tags": [
          "action-plans"
        ],
        "summary": "Get Plan",
        "operationId": "get_plan_api_action_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/action-plans/{plan_id}/confirm": {
      "post": {
        "tags": [
          "action-plans"
        ],
        "summary": "Confirm Plan",
        "operationId": "confirm_plan_api_action_plans__plan_id__confirm_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "unsubscribe_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unsubscribe Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/action-plans/{plan_id}/actions": {
      "get": {
        "tags": [
          "action-plans"
        ],
        "summary": "List Plan Actions",
        "operationId": "list_plan_actions_api_action_plans__plan_id__actions_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ActionListResponse"
      },
      "ActionPlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "digest": {
            "type": "string",
            "title": "Digest"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "digest",
          "confirmed",
          "items"
        ],
        "title": "ActionPlanResponse"
      },
      "ActionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "browser_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Browser Session Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "candidate_id",
          "method",
          "state"
        ],
        "title": "ActionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MailPreviewResponse": {
        "properties": {
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "recipient",
          "subject",
          "body"
        ],
        "title": "MailPreviewResponse"
      },
...
      "PlanConfirmRequest": {
        "properties": {
          "digest": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Digest"
          }
        },
        "type": "object",
        "required": [
          "digest"
        ],
        "title": "PlanConfirmRequest"
      },
      "PlanCreateRequest": {
        "properties": {
          "selections": {
            "items": {
              "$ref": "#/components/schemas/SelectionRequest"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Selections"
          }
        },
        "type": "object",
        "required": [
          "selections"
        ],
        "title": "PlanCreateRequest"
      },
      "PlanItemResponse": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "target_display": {
            "type": "string",
            "title": "Target Display"
          },
          "mail_preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MailPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "revision",
          "sender",
          "subject",
          "method",
          "target_display"
        ],
        "title": "PlanItemResponse"
      },
...
      "SelectionRequest": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "revision"
        ],
        "title": "SelectionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}